backend/requirements.txt
Flask==3.0.3
itsdangerous==2.1.2
Werkzeug==3.0.3
Jinja2==3.1.3
click==8.3.0

Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

python-dotenv==1.2.1

blinker==1.9.0
colorama==0.4.6
Pygments==2.19.2

MarkupSafe==3.0.3
packaging==25.0
typing_extensions==4.15.0

pytest==8.4.2
tomli==2.3.0
pluggy==1.6.0
iniconfig==2.3.0
exceptiongroup==1.3.0

requests==2.31.0
APScheduler==3.10.4
pytz==2024.1          
tzlocal==5.2         
